provenance writers now methods and not classes/algorithms
