+++ spinn_front_end_common/interface/interface_functions/front_end_common_interface_functions.xml
@@ -1317,60 +1317,6 @@
             <param_type>RouterProvenanceItems</param_type>
         </outputs>
     </algorithm>
-    <algorithm name="ProvenanceXMLWriter">
-        <python_module>spinn_front_end_common.interface.interface_functions</python_module>
-        <python_class>ProvenanceXMLWriter</python_class>
-        <input_definitions>
-            <parameter>
-                <param_name>provenance_data_items</param_name>
-                <param_type>ProvenanceItems</param_type>
-            </parameter>
-            <parameter>
-                <param_name>provenance_data_path</param_name>
-                <param_type>AppProvenanceFilePath</param_type>
-            </parameter>
-        </input_definitions>
-        <required_inputs>
-            <param_name>provenance_data_items</param_name>
-            <param_name>provenance_data_path</param_name>
-        </required_inputs>
-    </algorithm>
-    <algorithm name="ProvenanceJSONWriter">
-        <python_module>spinn_front_end_common.interface.interface_functions</python_module>
-        <python_class>ProvenanceJSONWriter</python_class>
-        <input_definitions>
-            <parameter>
-                <param_name>provenance_data_items</param_name>
-                <param_type>ProvenanceItems</param_type>
-            </parameter>
-            <parameter>
-                <param_name>provenance_data_path</param_name>
-                <param_type>AppProvenanceFilePath</param_type>
-            </parameter>
-        </input_definitions>
-        <required_inputs>
-            <param_name>provenance_data_items</param_name>
-            <param_name>provenance_data_path</param_name>
-        </required_inputs>
-    </algorithm>
-    <algorithm name="ProvenanceSQLWriter">
-        <python_module>spinn_front_end_common.interface.interface_functions</python_module>
-        <python_class>ProvenanceSQLWriter</python_class>
-        <input_definitions>
-            <parameter>
-                <param_name>provenance_data_items</param_name>
-                <param_type>ProvenanceItems</param_type>
-            </parameter>
-            <parameter>
-                <param_name>provenance_data_path</param_name>
-                <param_type>AppProvenanceFilePath</param_type>
-            </parameter>
-        </input_definitions>
-        <required_inputs>
-            <param_name>provenance_data_items</param_name>
-            <param_name>provenance_data_path</param_name>
-        </required_inputs>
-    </algorithm>
     <algorithm name="DatabaseInterface">
         <python_module>spinn_front_end_common.interface.interface_functions</python_module>
         <python_class>DatabaseInterface</python_class>
